requirements.txt
streamlit
pandas
matplotlib
supabase
bcrypt
openpyxl
python-dotenv
